Controolers must now be exposed manually. You can do that using the new method "expose(Class<?>...)". Pass your classes as parameters and you're done. Note that in order for your clients to be able to access your web methods you must define the required Beans.
Normal node require the "@Web" bean.
The 404 error node can be specified using the bean "@NotFound".
The method that specifies the "@Default" Bean will be the method that deals with every single file request, assuming the file exists.

Just like HttpController, WebSocketControllers can specify the "@Web" and "@NotFound" beans, but not the "@Default" bean.
The "@Default" beans does nothing for a WebSocket controller.
The WebSocketController methods must not specify a @Web bean at all, they are all scoped inside the class bean.