Added the "keepAlive" parameter in the config file. This parameter tells the server how long to wait for a any Thread to wait.
Changed the "thread" function in JavaScript.
The function will not create a simple Thread anymore, instead it will use the ThreadExecutor the core server also makes use of.
This means your callback is executed inside the manager ThreadPool created by the ThreadExecutor.
This also implies that you're not guaranteed your thread may or may not be queued depending if the Thread Pool is free or not at the moment of the call.