Exposing revamped a little
Exposing endpoints now allws parameters in the path string using @HttpService bean.

Exposing through CompleteAction is now encouraged as it requires less boilerplate code and the performance is overall identical since all these nodes are created beforehand (as the server starts), so all the runtime has to do is parse the bens and execute the HttpEvent.